requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.2.5
idna==2.8
idna-ssl==1.1.0
multidict==4.6.1
python-dotenv==0.10.3
typing-extensions==3.7.4.1
websockets==6.0
yarl==1.3.0
